error tweaks
